Various fixes and improvements

+++ src/mmh3.h
@@ -5,8 +5,8 @@
 
 __extension__ typedef unsigned __int128 uint128_t;
 
-void mmh3_x86_32(const void *key, int len, uint32_t seed, void *out);
-void mmh3_x86_128(const void *key, int len, uint32_t seed, void *out);
-void mmh3_x64_128(const void *key, int len, uint32_t seed, void *out);
+void mmh3_x86_32(const void *key, uint64_t len, uint32_t seed, void *out);
+void mmh3_x86_128(const void *key, uint64_t len, uint32_t seed, void *out);
+void mmh3_x64_128(const void *key, uint64_t len, uint32_t seed, void *out);
 
 #endif
